🔨 Switch Search to use useOvermind (#3048)
* 🔨 Switch Search to use useOvermind

* Use RouteComponentProps types

* Export Search as a named export

* Update search.ts
