+++ packages/app/src/app/pages/Search/index.tsx
@@ -3,33 +3,34 @@ import {
   ALGOLIA_APPLICATION_ID,
   ALGOLIA_DEFAULT_INDEX,
 } from '@codesandbox/common/lib/utils/config';
-import { Helmet } from 'react-helmet';
 import MaxWidth from '@codesandbox/common/lib/components/flex/MaxWidth';
 import Margin from '@codesandbox/common/lib/components/spacing/Margin';
+import { RouteComponentProps } from 'react-router-dom';
 import qs from 'qs';
-import React, { useCallback, useEffect, useRef, useState } from 'react';
+import React, {
+  FunctionComponent,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
+import { Helmet } from 'react-helmet';
 import {
-  InstantSearch,
-  SearchBox,
-  PoweredBy,
   Configure,
+  InstantSearch,
+  PoweredBy,
+  SearchBox,
 } from 'react-instantsearch/dom';
-import { History, Location } from 'history';
 
-import { Navigation } from 'app/pages/common/Navigation';
 import { useOvermind } from 'app/overmind';
+import { Navigation } from 'app/pages/common/Navigation';
 
 import 'instantsearch.css/themes/reset.css';
 
-import { Container, Content, Main, StyledTitle } from './elements';
+import { Container, Content, Main, Title } from './elements';
 import Filters from './Filters';
 import Results from './Results';
 import Styles from './search';
-
-interface ISearchProps {
-  history: History;
-  location: Location;
-}
 
 const updateAfter = 700;
 
@@ -38,11 +39,11 @@ const createURL = state => `?${qs.stringify(state)}`;
 const searchStateToUrl = (location, searchState) =>
   searchState ? `${location.pathname}${createURL(searchState)}` : '';
 
-const Search: React.FC<ISearchProps> = ({ history, location }) => {
+type Props = RouteComponentProps;
+export const Search: FunctionComponent<Props> = ({ history, location }) => {
   const {
     actions: { searchMounted },
   } = useOvermind();
-
   const [searchState, setSearchState] = useState(
     qs.parse(location.search.slice(1))
   );
@@ -83,6 +84,7 @@ const Search: React.FC<ISearchProps> = ({ history, location }) => {
       <Helmet>
         <title>Search - CodeSandbox</title>
       </Helmet>
+
       <Styles />
 
       <MaxWidth>
@@ -91,8 +93,8 @@ const Search: React.FC<ISearchProps> = ({ history, location }) => {
 
           <Content>
             <InstantSearch
-              appId={ALGOLIA_APPLICATION_ID}
               apiKey={ALGOLIA_API_KEY}
+              appId={ALGOLIA_APPLICATION_ID}
               createURL={createURL}
               indexName={ALGOLIA_DEFAULT_INDEX}
               onSearchStateChange={onSearchStateChange}
@@ -102,7 +104,7 @@ const Search: React.FC<ISearchProps> = ({ history, location }) => {
 
               <Main alignItems="flex-start">
                 <div>
-                  <StyledTitle>Search</StyledTitle>
+                  <Title>Search</Title>
 
                   <PoweredBy />
 
@@ -123,6 +125,3 @@ const Search: React.FC<ISearchProps> = ({ history, location }) => {
     </Container>
   );
 };
-
-// eslint-disable-next-line import/no-default-export
-export default Search;
